+++ Network/Messages/Message.cs
@@ -5,13 +5,13 @@ namespace Network.Messages
     [MessagePackObject]
     public class Message : IMessage
     {
-        public Message(long id, string typeMessage)
+        public Message(int id, string typeMessage)
         {
             Id = id;
             TypeMessage = typeMessage;
         }
         [Key(0)]
-        public long Id { get; set; }
+        public int Id { get; set; }
         [Key(1)]
         public string TypeMessage { get; set; }
     }
